requirements.txt
PySide6
ccxt
httpx
anyio
loguru
websocket-client
